requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-session==0.5.0
flask-wtf==1.2.1
flask-limiter==3.5.0
bcrypt==4.1.2
pyjwt==2.8.0
google-generativeai==0.3.2
reportlab==4.0.7
qrcode==7.4.2
Pillow==10.1.0
python-dotenv==1.0.0
